pyproject.toml
[project]
name = "bank-email-fetcher"
version = "0.1.0"
description = "Self-hosted bank transaction email fetcher and dashboard"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic-settings>=2.0,<3",
    "rich>=14.0,<15",
    "uvicorn[standard]>=0.34,<1",
    "jinja2>=3.1,<4",
    "sqlalchemy[asyncio]>=2.0,<3",
    "aiosqlite>=0.20,<1",
    "bank-email-parser",
    "bank-statement-parser",
    "cc-parser",
    "fastapi[standard]>=0.135,<1",
    "cryptography>=44.0,<47",
    "python-telegram-bot>=22.7,<23",
    "python-dateutil>=2.9.0.post0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.fastapi]
entrypoint = "bank_email_fetcher.main:create_app"

[tool.uv.sources]
bank-email-parser = { git = "https://github.com/akhilnarang/bank-email-parser" }
bank-statement-parser = { git = "https://github.com/akhilnarang/bank-statement-parser" }
cc-parser = { git = "https://github.com/akhilnarang/cc-parser" }

[dependency-groups]
dev = [
    "pytest>=9.0,<10",
    "pytest-anyio>=0.0.0",
]
